# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_kietle_zenreply_pkg_response.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  github_com_kietle_zenreply_pkg_response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_kietle_zenreply_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.APIError'
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Meta'
      success:
        type: boolean
    type: object
...
  handler.StartSessionRequest:
    properties:
      reason:
        example: Focused coding sprint
        maxLength: 500
        minLength: 1
        type: string
    required:
    - reason
    type: object
...
  model.DeepWorkSession:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: string
      is_active:
        description: IsActive is true when the session is currently ongoing.
        type: boolean
      reason:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  model.DeepWorkStatus:
    properties:
      is_active:
        type: boolean
      reason:
        type: string
      session_id:
        type: string
      started_at:
        type: string
      ttl_seconds:
        description: TTL is the remaining seconds until the session auto-expires (if
          set).
        type: integer
    type: object
  model.MessageLog:
    properties:
      auto_reply_text:
        type: string
      channel_id:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      id:
        type: string
      message_text:
        type: string
      original_ts:
        type: string
      sender_slack_id:
        type: string
      session_id:
        type: string
      thread_ts:
        type: string
      user_id:
        type: string
      was_sent:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@zenreply.app
    name: ZenReply Support
    url: https://zenreply.app/support
  description: ZenReply is an intelligent Slack auto-reply system for deep work sessions.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://zenreply.app/terms
  title: ZenReply API
  version: 1.0.0
paths:
  /deep-work/sessions:
    get:
      description: Returns a paginated list of all deep work sessions for the authenticated
        user
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.DeepWorkSession'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: List deep work sessions
      tags:
      - deep-work
    post:
      consumes:
      - application/json
      description: Begins a new deep work session for the authenticated user. Any
        existing active session is ended automatically.
      parameters:
      - description: Session details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.StartSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.DeepWorkSession'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Start a deep work session
      tags:
      - deep-work
  /deep-work/sessions/{id}:
    get:
      description: Returns details of a specific deep work session
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.DeepWorkSession'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get a specific session by ID
      tags:
      - deep-work
  /deep-work/sessions/active:
    delete:
      description: Terminates the currently active deep work session for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.DeepWorkSession'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: End the active deep work session
      tags:
      - deep-work
  /deep-work/status:
    get:
      description: Returns whether the authenticated user is currently in a deep work
        session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.DeepWorkStatus'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get current deep work status
      tags:
      - deep-work
...
  /logs:
    get:
      description: Returns a paginated list of all auto-reply messages sent by ZenReply
        for the authenticated user
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.MessageLog'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: List auto-reply message logs
      tags:
      - logs
  /logs/sessions/{session_id}:
    get:
      description: Returns all auto-reply messages sent during a specific deep work
        session
      parameters:
      - description: Session UUID
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.MessageLog'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: List message logs for a specific session
      tags:
      - logs
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <your-jwt-token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"